requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy~=1.16.2
Pillow~=5.4.1
scipy>=0.17.0

requests~=2.21.0
flask~=1.0.2
pymysql~=0.9.3
scikit-learn~=0.20.3
setuptools~=40.8.0
